:bookmark: Release

+++ commands/command-new/src/commands/new.stage.ts
@@ -3,7 +3,6 @@ import {
   checkForProjectDirectory,
   CliArgs,
   CliFlags,
-  flags,
   Log,
   MarketplacePlugin,
   PluginCommand,
@@ -23,32 +22,20 @@ import latestVersion from 'latest-version';
 import _merge from 'lodash.merge';
 import { join as joinPaths, resolve } from 'path';
 import { Choice } from 'prompts';
-import { promptPlugins, promptServer } from '../Prompts';
+import { promptPlugins, promptServer } from '../prompts';
 import { fetchMarketPlace, insert, runNpmInstall } from '../utilities';
 
-export type NewStageArgs = CliArgs<typeof NewStage>;
-export type NewStageFlags = CliFlags<typeof NewStage>;
-
 export interface NewStageContext extends PluginContext {
-  args: NewStageArgs;
-  flags: NewStageFlags;
+  args: CliArgs<typeof NewStage>;
+  flags: CliFlags<typeof NewStage>;
 }
 
 export type NewStageEvents = 'before.new:stage' | 'new:stage' | 'after.new:stage';
 
 export class NewStage extends PluginCommand<NewStageEvents> {
   static id = 'new:stage';
-  // Prints out a description for this command.
   static description = 'Creates a new stage.';
-  // Prints out examples for this command.
   static examples: string[] = [];
-  // Defines flags for this command, such as --help.
-  static flags = {
-    overwrite: flags.boolean({
-      description: 'Forces overwriting an existing project.',
-    }),
-    ...PluginCommand.flags,
-  };
   // Defines arguments that can be passed to the command.
   static args = [
     <const>{
@@ -72,7 +59,7 @@ export class NewStage extends PluginCommand<NewStageEvents> {
   async checkForExistingStage(): Promise<void> {
     const appFileName = `app.${this.$context.args.stage}.ts`;
 
-    if (existsSync(joinPaths('src', appFileName)) && !this.$context.flags.overwrite) {
+    if (existsSync(joinPaths('src', appFileName))) {
       const { overwrite } = await promptOverwrite(
         `Stage ${printHighlight(
           this.$context.args.stage,
@@ -100,7 +87,10 @@ export class NewStage extends PluginCommand<NewStageEvents> {
         description: plugin.description,
       }));
     const { server } = await promptServer(servers);
-    const serverFileName = `server.${server.module.toLowerCase()}`;
+
+    const serverFileName: string | undefined = server
+      ? `server.${server.module.toLowerCase()}`
+      : undefined;
 
     // Offer the user a range of plugins consisting of database and analytics plugins.
     const availableAppPlugins: prompt.Choice[] = marketPlacePlugins
@@ -159,20 +149,20 @@ export class NewStage extends PluginCommand<NewStageEvents> {
         packageJson.dependencies[plugin.package] = await latestVersion(plugin.package);
       }
       // Add selected server dependency to package.json
-      packageJson.dependencies[server.package] = await latestVersion(server.package);
+      if (server) {
+        packageJson.dependencies[server.package] = await latestVersion(server.package);
+      }
 
       // Create new npm scripts
       const appStartPath: string[] = [`app.${this.$context.args.stage}.js`];
-      if (this.$cli.$project!.isTypeScriptProject()) {
+      if (this.$cli.project!.isTypeScriptProject()) {
         appStartPath.unshift('dist');
       } else {
         appStartPath.unshift('src');
       }
       const appBundlePath: string = joinPaths(
         'src',
-        `app.${this.$context.args.stage}.${
-          this.$cli.$project!.isTypeScriptProject() ? 'ts' : 'js'
-        }`,
+        `app.${this.$context.args.stage}.${this.$cli.project!.isTypeScriptProject() ? 'ts' : 'js'}`,
       );
 
       packageJson.scripts[
@@ -184,16 +174,34 @@ export class NewStage extends PluginCommand<NewStageEvents> {
 
       writeFileSync('package.json', JSON.stringify(packageJson, null, 2));
       await runNpmInstall('./');
-
-      copyFileSync(
-        joinPaths('node_modules', server.package, 'boilerplate', `${serverFileName}.ts`),
-        joinPaths('src', `${serverFileName}.ts`),
-      );
     });
 
     stageTask.add(addPluginsTask, installTask);
 
     await stageTask.run();
+
+    const serverFilePath: string | undefined = server
+      ? joinPaths('src', `${serverFileName}.ts`)
+      : undefined;
+    Log.verbose(serverFilePath!);
+
+    if (serverFilePath) {
+      if (existsSync(serverFilePath)) {
+        Log.spacer();
+        const { overwrite } = await promptOverwrite(
+          `${serverFilePath} already exists. Do you want to overwrite it?`,
+        );
+
+        if (overwrite === ANSWER_CANCEL) {
+          return;
+        }
+      }
+
+      copyFileSync(
+        joinPaths('node_modules', server!.package, 'boilerplate', `${serverFileName}.ts`),
+        serverFilePath,
+      );
+    }
   }
 
   async run(): Promise<void> {
@@ -204,7 +212,7 @@ export class NewStage extends PluginCommand<NewStageEvents> {
     Log.info(printSubHeadline('Learn more: https://jovo.tech/docs/cli/new:stage'));
     Log.spacer();
 
-    const { args, flags }: { args: NewStageArgs; flags: NewStageFlags } = this.parse(NewStage);
+    const { args, flags } = this.parse(NewStage);
 
     _merge(this.$context, { args, flags });
